+++ include/Coordinates.hpp
@@ -4,10 +4,10 @@
 #include <cstdint>
 
 struct Coordinates {
-    uint8_t x;
-    uint8_t y;
+    int16_t x;
+    int16_t y;
 
-    Coordinates(uint8_t x, uint8_t y) : x(x), y(y) {}
+    Coordinates(int16_t x, int16_t y) : x(x), y(y) {}
 
     Coordinates() = default;
 };
